fix: jsandoval-2025365 ha solucionado varios errores de sintaxis y con importes en los archivos UsuarioController, Usuario, IUsuarioRepository y UsuarioService

+++ src/main/java/com/justinsandoval/entity/Usuario.java
@@ -14,7 +14,7 @@ public class Usuario {
     //Atributos
     @Id
     @Column(name = "codigo_usuario")
-    private int co_usuario;
+    private long co_usuario;
     @Column
     private String username;
     @Column
