+++ src/main/java/com/justinsandoval/repository/IUsuarioRepository.java
@@ -3,6 +3,6 @@ package com.justinsandoval.repository;
 import com.justinsandoval.entity.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IUsuarioRepository extends JpaRepository<Usuario , int> {
+public interface IUsuarioRepository extends JpaRepository<Usuario , Long> {
 }
